Refactor `starlite.handlers.http_handlers`. (#1244)
